pyproject.toml
[project]
name = "nats-headers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.9.0",
]
